feat(eslint): implement custom ESLint rules for Citron guidelines
- Added a new ESLint plugin with custom rules for file naming conventions, helper class structure, DTO restrictions, repository return types, and service statelessness.
- Removed old rule implementations and consolidated them into a single file for better organization.
- Updated the main ESLint configuration to reflect the new plugin structure.
